# Fix serious performance degradation when building stack traces