--- a/pkg/dwarf/reader/reader.go
+++ b/pkg/dwarf/reader/reader.go
@@ -426,6 +426,9 @@ childLoop:
 		default:
 			irdr.reader.SkipChildren()
 		}
+		if rentry == nil {
+			break
+		}
 	}
 
 	if rentry != nil && rentry.Tag == dwarf.TagInlinedSubroutine {