--- a/pkg/proc/registers.go
+++ b/pkg/proc/registers.go
@@ -18,7 +18,6 @@ type Registers interface {
 	PC() uint64
 	SP() uint64
 	BP() uint64
-	CX() uint64
 	TLS() uint64
 	// GAddr returns the address of the G variable if it is known, 0 and false otherwise
 	GAddr() (uint64, bool)