@aarzilli 2020-01-27
This PR adds a benchmark for conditional breakpoints and adds a serious performance degradation to stacktraces introduced by commit adb1746c6019cd2ac40e37bb00d71647245268a2.

This part of a [larger series of optimizations](https://github.com/go-delve/delve/pull/1846) that reduces the latency of a conditional breakpoint evaluation from 8ms to 0.9ms. The majority of the latency is caused by this particular bug (from 8ms down to 2.2ms).